🚀 [Feat]: Add logic to house-appeal input field & Separate template
- 입력 후 enter시에, onChange로 입력값을 받아 appeal state에 저장해 house_appeal 배열에 추가하고 badge를 생성하도록 함
- 생성된 badge를 클릭하면 house_appeal에서 해당 값이 삭제되도록 함
- 차후 validation등 추가를 고려해 form template를 분리함
